+++ src/__tests__/ErrorPage.test.jsx
@@ -5,7 +5,15 @@ import ErrorPage from "../components/ErrorPage";
 describe("ErrorPage component", () => {
   it("The ErrorPage component is rendered", () => {
     render(<ErrorPage />);
-    const errorPageSection = screen.getByText("Error");
+    // Why can't I use getByRole('region')?
+    // const errorPageSection = screen.getByText("Error");
+    const errorPageSection = screen.getByRole("region");
     expect(errorPageSection).toBeInTheDocument();
+  });
+
+  it('The ErrorPage component text is "404 Page not found"', () => {
+    const errorMessage = "404 Page not found";
+    render(<ErrorPage errorMessage={errorMessage} />);
+    expect(screen.getByText(errorMessage)).toHaveTextContent(errorMessage);
   });
 });