+++ src/layouts/ProductLayout.jsx
@@ -1,0 +1,11 @@
+import { Outlet } from "react-router-dom";
+
+const ProductLayout = () => {
+  return (
+    <dialog>
+      <Outlet />
+    </dialog>
+  );
+};
+
+export default ProductLayout;